server/requirements.txt
django
djangorestframework
pre-commit

-e ../lib/clp_common/
